requirements.txt
fastapi
uvicorn
tensorflow
pillow
mangum
